+++ src/plugins/kibana/public/settings/sections/kibi_dashboard_groups/services/_saved_dashboard_group.js
@@ -24,7 +24,7 @@ define(function (require) {
           iconCss: 'string',
           iconUrl: 'string',
           hide: 'boolean',
-          version: 'long'
+          version: 'integer'
         },
 
         defaults: {
